+++ R/R6Class-BipartiteSBM_fit.R
@@ -58,7 +58,7 @@ BipartiteSBM_fit <-
         args <- c(args, blockmodelsOptions)
 
         ## model construction
-        model_type <- ifelse(self$nbCovariates > 0, paste0(model,"_covariates"), private$model)
+        model_type <- ifelse(self$nbCovariates > 0, paste0(private$model,"_covariates"), private$model)
         if (model_type == 'bernoulli_covariates' & fast == TRUE) model_type <- 'bernoulli_covariates_fast'
         private$BMobject <- do.call(paste0("BM_", model_type), args)
 
@@ -75,8 +75,8 @@ BipartiteSBM_fit <-
       predict = function(covarList = self$covarList) {
         stopifnot(is.list(covarList), self$nbCovariates == length(covarList))
         if (length(covarList) > 0) {
-          stopifnot(all.equal(self$dimension[1], sapply(covarList, nrow)),
-                    all.equal(self$dimension[2], sapply(covarList, ncol)))
+          stopifnot(all(sapply(covarList, nrow) == self$dimension[1]),
+                    all(sapply(covarList, ncol) == self$dimension[2]))
         }
         mu <- private$tau[[1]] %*% private$theta$mean %*% t(private$tau[[2]])
         if (length(self$covList) > 0) mu <- private$invlink(private$link(mu) + self$covarEffect)