+++ tests/testthat/test-SimpleSBM_fit_covariates.R
@@ -3,24 +3,33 @@ set.seed(1234)
 rmse <- function(theta, theta_star) { sqrt(sum((theta - theta_star)^2)/sum(theta_star^2)) }
 
 ## Common parameters
-nbNodes  <- 90
-nbBlocks <- 3
-blockProp <- c(.5, .25, .25) # group proportions
+nbNodes  <- 40
+nbBlocks <- 2
+blockProp <- c(.5, .5) # group proportions
+covarParam <- c(-2,2)
+covar1 <- matrix(rnorm(nbNodes**2), nbNodes, nbNodes)
+covar2 <- matrix(rnorm(nbNodes**2), nbNodes, nbNodes)
+covarList_directed <- list(covar1 = covar1, covar2 = covar2)
 
-test_that("SimpleSBM_fit 'Bernoulli' model, undirected, no covariate", {
+covar1 <- covar1 + t(covar1)
+covar2 <- covar2 + t(covar2)
+covarList <- list(covar1 = covar1, covar2 = covar2)
+
+test_that("SimpleSBM_fit 'Bernoulli' model, undirected, one covariate", {
 
   ## SIMPLE UNDIRECTED BERNOULLI SBM
-  means <- diag(.4, 3) + 0.05
+  means <- diag(.4, 2) + 0.05
   connectParam <- list(mean = means)
 
   ## Basic construction - check for wrong specifications
-  mySampler <- SimpleSBM_sampler$new('bernoulli', nbNodes, FALSE, blockProp, connectParam)
+  mySampler <- SimpleSBM_sampler$new('bernoulli', nbNodes, FALSE, blockProp, connectParam, covarParam[1], covarList[1])
 
   ## Construction----------------------------------------------------------------
-  mySBM <- SimpleSBM_fit$new(mySampler$netMatrix, 'bernoulli', FALSE)
-  expect_error(SimpleSBM_fit$new(mySampler$netMatrix, 'bernouilli', FALSE))
-  expect_error(SimpleSBM_fit$new(mySampler$netMatrix[1:20, 1:30], 'bernouilli', FALSE))
-  expect_error(SimpleSBM_fit$new(mySampler$netMatrix, 'bernoulli', TRUE))
+  mySBM <- SimpleSBM_fit$new(mySampler$netMatrix, 'bernoulli', FALSE, covarList[1])
+  expect_error(SimpleSBM_fit$new(mySampler$netMatrix, 'bernouilli', FALSE, covarList[1]))
+  expect_error(SimpleSBM_fit$new(mySampler$netMatrix[1:20, 1:30], 'bernouilli', FALSE, covarList[1]))
+  expect_error(SimpleSBM_fit$new(mySampler$netMatrix, 'bernoulli', TRUE, covarList[1]))
+  expect_error(SimpleSBM_fit$new(mySampler$netMatrix, 'bernoulli', FALSE, covarList[[1]]))
 
   ## Checking class
   expect_true(inherits(mySBM, "SBM"))
@@ -40,9 +49,9 @@ test_that("SimpleSBM_fit 'Bernoulli' model, undirected, no covariate", {
 
   ## covariates
   expect_null(mySBM$covarExpect)
-  expect_equal(mySBM$nbCovariates, 0)
-  expect_equal(mySBM$covarList, list())
-  expect_equal(mySBM$covarParam, numeric(0))
+  expect_equal(mySBM$nbCovariates, 1)
+  expect_equal(mySBM$covarList, covarList[1])
+  expect_equal(mySBM$covarParam, c(0))
 
   ## S3 methods
   expect_equal(coef(mySBM, 'connectivity'), mySBM$connectParam)
@@ -50,14 +59,20 @@ test_that("SimpleSBM_fit 'Bernoulli' model, undirected, no covariate", {
   expect_equal(coef(mySBM, 'covariates')  , mySBM$covarParam)
 
   ## Estimation-----------------------------------------------------------------
-  mySBM$optimize(verbosity = 0)
-  mySBM$setModel(3)
+  mySBM$optimize(verbosity = 0, fast = TRUE)
+  mySBM$setModel(2)
 
   ## Expectation
   expect_equal(dim(mySBM$expectation), c(nbNodes, nbNodes))
   expect_true(all(mySBM$expectation >= 0, na.rm = TRUE))
   expect_true(all(mySBM$expectation <= 1, na.rm = TRUE))
   expect_null(mySBM$connectParam$var)
+
+  ## blocks
+  expect_equal(mySBM$nbBlocks, nbBlocks)
+  expect_equal(dim(mySBM$probMemberships), c(nbNodes, nbBlocks))
+  expect_equal(sort(unique(mySBM$memberships)), 1:nbBlocks)
+  expect_equal(length(mySBM$memberships), nbNodes)
 
   ## S3 methods
   expect_equal(coef(mySBM, 'connectivity'), mySBM$connectParam)
@@ -66,33 +81,25 @@ test_that("SimpleSBM_fit 'Bernoulli' model, undirected, no covariate", {
   expect_equal(mySBM$predict(), predict(mySBM))
   expect_equal(mySBM$fitted, fitted(mySBM))
   expect_equal(mySBM$fitted, predict(mySBM))
-
-  ## blocks
-  expect_equal(mySBM$nbBlocks, nbBlocks)
-  expect_equal(dim(mySBM$probMemberships), c(nbNodes, nbBlocks))
-  expect_equal(sort(unique(mySBM$memberships)), 1:nbBlocks)
-  expect_equal(length(mySBM$memberships), nbNodes)
-
-  ## correctness
-  expect_lt(rmse(mySBM$connectParam$mean, means), 0.25)
-  expect_lt(1 - aricode::ARI(mySBM$memberships, mySampler$memberships), 0.25)
+  expect_equal(predict(mySBM, covarList[1]), fitted(mySBM))
+  expect_error(predict(mySBM, covarList))
 
 })
 
-test_that("SimpleSBM_fit 'Bernoulli' model, directed, no covariate", {
+test_that("SimpleSBM_fit 'Bernoulli' model, directed, one covariate", {
 
   ## SIMPLE UNDIRECTED BERNOULLI SBM
-  means <- matrix(c(1:9)/10, 3,  3)
+  means <- matrix(c(0.1, 0.4, 0.6, 0.9), 2,  2)
   connectParam <- list(mean = means)
 
   ## Basic construction - check for wrong specifications
-  mySampler <- SimpleSBM_sampler$new('bernoulli', nbNodes, TRUE, blockProp, connectParam)
+  mySampler <- SimpleSBM_sampler$new('bernoulli', nbNodes, TRUE, blockProp, connectParam, covarParam[1], covarList_directed[1])
 
   ## Construction----------------------------------------------------------------
-  mySBM <- SimpleSBM_fit$new(mySampler$netMatrix, 'bernoulli', TRUE)
-  expect_error(SimpleSBM_fit$new(mySampler$netMatrix, 'bernouilli', TRUE))
-  expect_error(SimpleSBM_fit$new(mySampler$netMatrix, 'bernoulli', FALSE))
-  expect_error(SimpleSBM_fit$new(mySampler$netMatrix[1:20, 1:30], 'bernouilli', FALSE))
+  mySBM <- SimpleSBM_fit$new(mySampler$netMatrix, 'bernoulli', TRUE, covarList_directed[1])
+  expect_error(SimpleSBM_fit$new(mySampler$netMatrix, 'bernouilli', TRUE, covarList_directed[1]))
+  expect_error(SimpleSBM_fit$new(mySampler$netMatrix, 'bernoulli', FALSE, covarList_directed[1]))
+  expect_error(SimpleSBM_fit$new(mySampler$netMatrix[1:20, 1:30], 'bernouilli', FALSE, covarList_directed[1]))
 
   ## Checking class
   expect_true(inherits(mySBM, "SBM"))
@@ -112,9 +119,9 @@ test_that("SimpleSBM_fit 'Bernoulli' model, directed, no covariate", {
 
   ## covariates
   expect_null(mySBM$covarExpect)
-  expect_equal(mySBM$nbCovariates, 0)
-  expect_equal(mySBM$covarList, list())
-  expect_equal(mySBM$covarParam, numeric(0))
+  expect_equal(mySBM$nbCovariates, 1)
+  expect_equal(mySBM$covarList, covarList_directed[1])
+  expect_equal(mySBM$covarParam, c(0))
 
   ## S3 methods
   expect_equal(coef(mySBM, 'connectivity'), mySBM$connectParam)
@@ -122,8 +129,8 @@ test_that("SimpleSBM_fit 'Bernoulli' model, directed, no covariate", {
   expect_equal(coef(mySBM, 'covariates')  , mySBM$covarParam)
 
   ## Estimation-----------------------------------------------------------------
-  mySBM$optimize(verbosity = 0)
-  mySBM$setModel(3)
+  mySBM$optimize(verbosity = 0, fast = TRUE)
+  mySBM$setModel(2)
 
   ## Expectation
   expect_equal(dim(mySBM$expectation), c(nbNodes, nbNodes))
@@ -144,27 +151,24 @@ test_that("SimpleSBM_fit 'Bernoulli' model, directed, no covariate", {
   expect_equal(mySBM$predict(), predict(mySBM))
   expect_equal(mySBM$fitted, fitted(mySBM))
   expect_equal(mySBM$fitted, predict(mySBM))
-
-  ## correctness
-  expect_lt(rmse(sort(mySBM$connectParam$mean), means), 0.2)
-  expect_lt(1 - aricode::ARI(mySBM$memberships, mySampler$memberships), 0.2)
+  expect_equal(predict(mySBM, covarList[1]), fitted(mySBM))
 
 })
 
-test_that("SimpleSBM_fit 'Poisson' model, undirected, no covariate", {
+test_that("SimpleSBM_fit 'Poisson' model, undirected, two covariates", {
 
   ## SIMPLE UNDIRECTED POISSON SBM
-  means <- diag(15., 3) + 5
+  means <- diag(15, 2) + 5
   connectParam <- list(mean = means)
 
   ## Basic construction - check for wrong specifications
-  mySampler <- SimpleSBM_sampler$new('poisson', nbNodes, FALSE, blockProp, connectParam)
+  mySampler <- SimpleSBM_sampler$new('poisson', nbNodes, FALSE, blockProp, connectParam, covarParam[1], covarList[1])
 
   ## Construction----------------------------------------------------------------
-  mySBM <- SimpleSBM_fit$new(mySampler$netMatrix, 'poisson', FALSE)
-  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix, 'poison', FALSE))
-  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix[1:20, 1:30], 'poisson', FALSE))
-  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix, 'poisson', TRUE))
+  mySBM <- SimpleSBM_fit$new(mySampler$netMatrix, 'poisson', FALSE, covarList[1])
+  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix, 'poison', FALSE, covarList[1]))
+  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix[1:20, 1:30], 'poisson', FALSE, covarList[1]))
+  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix, 'poisson', TRUE, covarList[1]))
 
   ## Checking class
   expect_true(inherits(mySBM, "SBM"))
@@ -184,9 +188,9 @@ test_that("SimpleSBM_fit 'Poisson' model, undirected, no covariate", {
 
   ## covariates
   expect_null(mySBM$covarExpect)
-  expect_equal(mySBM$nbCovariates, 0)
-  expect_equal(mySBM$covarList, list())
-  expect_equal(mySBM$covarParam, numeric(0))
+  expect_equal(mySBM$nbCovariates, 1)
+  expect_equal(mySBM$covarList, covarList[1])
+  expect_equal(mySBM$covarParam, c(0))
 
   ## S3 methods
   expect_equal(coef(mySBM, 'connectivity'), mySBM$connectParam)
@@ -195,12 +199,18 @@ test_that("SimpleSBM_fit 'Poisson' model, undirected, no covariate", {
 
   ## Estimation-----------------------------------------------------------------
   mySBM$optimize(verbosity = 0)
-  mySBM$setModel(3)
+  mySBM$setModel(2)
 
   ## Expectation
   expect_equal(dim(mySBM$expectation), c(nbNodes, nbNodes))
   expect_true(all(mySampler$expectation >= 0, na.rm = TRUE))
   expect_null(mySBM$connectParam$var)
+
+  ## blocks
+  expect_equal(mySBM$nbBlocks, nbBlocks)
+  expect_equal(dim(mySBM$probMemberships), c(nbNodes, nbBlocks))
+  expect_equal(sort(unique(mySBM$memberships)), 1:nbBlocks)
+  expect_equal(length(mySBM$memberships), nbNodes)
 
   ## S3 methods
   expect_equal(coef(mySBM, 'connectivity'), mySBM$connectParam)
@@ -209,33 +219,24 @@ test_that("SimpleSBM_fit 'Poisson' model, undirected, no covariate", {
   expect_equal(mySBM$predict(), predict(mySBM))
   expect_equal(mySBM$fitted, fitted(mySBM))
   expect_equal(mySBM$fitted, predict(mySBM))
-
-  ## blocks
-  expect_equal(mySBM$nbBlocks, nbBlocks)
-  expect_equal(dim(mySBM$probMemberships), c(nbNodes, nbBlocks))
-  expect_equal(sort(unique(mySBM$memberships)), 1:nbBlocks)
-  expect_equal(length(mySBM$memberships), nbNodes)
-
-  ## correctness
-  expect_lt(rmse(mySBM$connectParam$mean, means), 1e-1)
-  expect_lt(1 - aricode::ARI(mySBM$memberships, mySampler$memberships), 1e-1)
+  expect_equal(predict(mySBM, covarList[1]), fitted(mySBM))
 
 })
 
-test_that("SimpleSBM_fit 'Poisson' model, directed, no covariate", {
+test_that("SimpleSBM_fit 'Poisson' model, directed, two covariates", {
 
   ## SIMPLE UNDIRECTED BERNOULLI SBM
-  means <- matrix(1:9, 3,  3)
+  means <- matrix(c(1, 4, 7, 9), 2,  2)
   connectParam <- list(mean = means)
 
   ## Basic construction - check for wrong specifications
-  mySampler <- SimpleSBM_sampler$new('poisson', nbNodes, TRUE, blockProp, connectParam)
+  mySampler <- SimpleSBM_sampler$new('poisson', nbNodes, TRUE, blockProp, connectParam, covarParam[1], covarList_directed[1])
 
   ## Construction----------------------------------------------------------------
-  mySBM <- SimpleSBM_fit$new(mySampler$netMatrix, 'poisson', TRUE)
-  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix, 'poison', TRUE))
-  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix[1:20, 1:30], 'poisson', FALSE))
-  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix, 'poisson', FALSE))
+  mySBM <- SimpleSBM_fit$new(mySampler$netMatrix, 'poisson', TRUE, covarList_directed[1])
+  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix, 'poison', TRUE, covarList_directed[1]))
+  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix[1:20, 1:30], 'poisson', FALSE, covarList_directed[1]))
+  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix, 'poisson', FALSE, covarList_directed[1]))
 
   ## Checking class
   expect_true(inherits(mySBM, "SBM"))
@@ -255,9 +256,9 @@ test_that("SimpleSBM_fit 'Poisson' model, directed, no covariate", {
 
   ## covariates
   expect_null(mySBM$covarExpect)
-  expect_equal(mySBM$nbCovariates, 0)
-  expect_equal(mySBM$covarList, list())
-  expect_equal(mySBM$covarParam, numeric(0))
+  expect_equal(mySBM$nbCovariates, 1)
+  expect_equal(mySBM$covarList, covarList_directed[1])
+  expect_equal(mySBM$covarParam, c(0))
 
   ## S3 methods
   expect_equal(coef(mySBM, 'connectivity'), mySBM$connectParam)
@@ -266,7 +267,7 @@ test_that("SimpleSBM_fit 'Poisson' model, directed, no covariate", {
 
   ## Estimation-----------------------------------------------------------------
   mySBM$optimize(verbosity = 0)
-  mySBM$setModel(3)
+  mySBM$setModel(2)
 
   ## Expectation
   expect_equal(dim(mySBM$expectation), c(nbNodes, nbNodes))
@@ -286,28 +287,25 @@ test_that("SimpleSBM_fit 'Poisson' model, directed, no covariate", {
   expect_equal(mySBM$predict(), predict(mySBM))
   expect_equal(mySBM$fitted, fitted(mySBM))
   expect_equal(mySBM$fitted, predict(mySBM))
-
-  ## correctness
-  expect_lt(rmse(sort(mySBM$connectParam$mean), means), 1e-1)
-  expect_lt(1 - aricode::ARI(mySBM$memberships, mySampler$memberships), 1e-1)
+  expect_equal(predict(mySBM, covarList[1]), fitted(mySBM))
 
 })
 
 
-test_that("SimpleSBM_fit 'Gaussian' model, undirected, no covariate", {
+test_that("SimpleSBM_fit 'Gaussian' model, undirected, two covariates", {
 
   ## SIMPLE UNDIRECTED GAUSSIAN SBM
-  means <- diag(15., 3) + 5 # connectivity matrix: affiliation network
+  means <- diag(15., 2) + 5 # connectivity matrix: affiliation network
   connectParam <- list(mean = means, var = 2)
 
   ## Basic construction - check for wrong specifications
-  mySampler <- SimpleSBM_sampler$new('gaussian', nbNodes, FALSE, blockProp, connectParam)
+  mySampler <- SimpleSBM_sampler$new('gaussian', nbNodes, FALSE, blockProp, connectParam, covarParam, covarList)
 
   ## Construction----------------------------------------------------------------
-  mySBM <- SimpleSBM_fit$new(mySampler$netMatrix, 'gaussian', FALSE)
-  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix, 'normal', FALSE))
-  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix[1:20, 1:30], 'gaussian', FALSE))
-  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix, 'gaussian', TRUE))
+  mySBM <- SimpleSBM_fit$new(mySampler$netMatrix, 'gaussian', FALSE, covarList)
+  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix, 'normal', FALSE, covarList))
+  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix[1:20, 1:30], 'gaussian', FALSE, covarList))
+  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix, 'gaussian', TRUE, covarList))
 
   ## Checking class
   expect_true(inherits(mySBM, "SBM"))
@@ -327,9 +325,9 @@ test_that("SimpleSBM_fit 'Gaussian' model, undirected, no covariate", {
 
   ## covariates
   expect_null(mySBM$covarExpect)
-  expect_equal(mySBM$nbCovariates, 0)
-  expect_equal(mySBM$covarList, list())
-  expect_equal(mySBM$covarParam, numeric(0))
+  expect_equal(mySBM$nbCovariates, 2)
+  expect_equal(mySBM$covarList, covarList)
+  expect_equal(mySBM$covarParam, c(0,0))
 
   ## S3 methods
   expect_equal(coef(mySBM, 'connectivity'), mySBM$connectParam)
@@ -338,7 +336,7 @@ test_that("SimpleSBM_fit 'Gaussian' model, undirected, no covariate", {
 
   ## Estimation-----------------------------------------------------------------
   mySBM$optimize(verbosity = 0)
-  mySBM$setModel(3)
+  mySBM$setModel(2)
 
   ## Expectation
   expect_equal(dim(mySBM$expectation), c(nbNodes, nbNodes))
@@ -350,6 +348,11 @@ test_that("SimpleSBM_fit 'Gaussian' model, undirected, no covariate", {
   expect_equal(sort(unique(mySBM$memberships)), 1:nbBlocks)
   expect_equal(length(mySBM$memberships), nbNodes)
 
+  ## correctness
+  expect_lt(rmse(mySBM$connectParam$mean, means), 1e-1)
+  expect_lt(rmse(mySBM$covarParam, covarParam), 0.1)
+  expect_lt(1 - aricode::ARI(mySBM$memberships, mySampler$memberships), 1e-1)
+
   ## S3 methods
   expect_equal(coef(mySBM, 'connectivity'), mySBM$connectParam)
   expect_equal(coef(mySBM, 'block')       , mySBM$blockProp)
@@ -357,27 +360,25 @@ test_that("SimpleSBM_fit 'Gaussian' model, undirected, no covariate", {
   expect_equal(mySBM$predict(), predict(mySBM))
   expect_equal(mySBM$fitted, fitted(mySBM))
   expect_equal(mySBM$fitted, predict(mySBM))
-
-  ## correctness
-  expect_lt(rmse(mySBM$connectParam$mean, means), 1e-1)
-  expect_lt(1 - aricode::ARI(mySBM$memberships, mySampler$memberships), 1e-1)
+  expect_equal(predict(mySBM, covarList), fitted(mySBM))
 
 })
 
-test_that("SimpleSBM_fit 'Gaussian' model, undirected, no covariate", {
+
+test_that("SimpleSBM_fit 'Gaussian' model, undirected, two covariates", {
 
   ## SIMPLE UNDIRECTED GAUSSIAN SBM
-  means <- matrix(1:9,3,3)
+  means <- matrix(c(1, 4, 7, 10),2,2)
   connectParam <- list(mean = means, var = 2)
 
   ## Basic construction - check for wrong specifications
-  mySampler <- SimpleSBM_sampler$new('gaussian', nbNodes, TRUE, blockProp, connectParam)
+  mySampler <- SimpleSBM_sampler$new('gaussian', nbNodes, TRUE, blockProp, connectParam, covarParam, covarList_directed)
 
   ## Construction----------------------------------------------------------------
-  mySBM <- SimpleSBM_fit$new(mySampler$netMatrix, 'gaussian', TRUE)
-  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix, 'normal', TRUE))
-  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix[1:20, 1:30], 'gaussian', TRUE))
-  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix, 'gaussian', FALSE))
+  mySBM <- SimpleSBM_fit$new(mySampler$netMatrix, 'gaussian', TRUE, covarList_directed)
+  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix, 'normal', TRUE, covarList_directed))
+  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix[1:20, 1:30], 'gaussian', TRUE, covarList_directed))
+  expect_error(SimpleSBM_fit$new(SamplerBernoulli$netMatrix, 'gaussian', FALSE, covarList_directed))
 
   ## Checking class
   expect_true(inherits(mySBM, "SBM"))
@@ -397,9 +398,9 @@ test_that("SimpleSBM_fit 'Gaussian' model, undirected, no covariate", {
 
   ## covariates
   expect_null(mySBM$covarExpect)
-  expect_equal(mySBM$nbCovariates, 0)
-  expect_equal(mySBM$covarList, list())
-  expect_equal(mySBM$covarParam, numeric(0))
+  expect_equal(mySBM$nbCovariates, 2)
+  expect_equal(mySBM$covarList, covarList_directed)
+  expect_equal(mySBM$covarParam, c(0,0))
 
   ## S3 methods
   expect_equal(coef(mySBM, 'connectivity'), mySBM$connectParam)
@@ -408,7 +409,7 @@ test_that("SimpleSBM_fit 'Gaussian' model, undirected, no covariate", {
 
   ## Estimation-----------------------------------------------------------------
   mySBM$optimize(verbosity = 0)
-  mySBM$setModel(3)
+  mySBM$setModel(2)
 
   ## Expectation
   expect_equal(dim(mySBM$expectation), c(nbNodes, nbNodes))
@@ -420,6 +421,11 @@ test_that("SimpleSBM_fit 'Gaussian' model, undirected, no covariate", {
   expect_equal(sort(unique(mySBM$memberships)), 1:nbBlocks)
   expect_equal(length(mySBM$memberships), nbNodes)
 
+  ## correctness
+  expect_lt(rmse(sort(mySBM$connectParam$mean), means), 1e-1)
+  expect_lt(rmse(mySBM$covarParam, covarParam), 0.1)
+  expect_lt(1 - aricode::ARI(mySBM$memberships, mySampler$memberships), 1e-1)
+
   ## S3 methods
   expect_equal(coef(mySBM, 'connectivity'), mySBM$connectParam)
   expect_equal(coef(mySBM, 'block')       , mySBM$blockProp)
@@ -427,9 +433,6 @@ test_that("SimpleSBM_fit 'Gaussian' model, undirected, no covariate", {
   expect_equal(mySBM$predict(), predict(mySBM))
   expect_equal(mySBM$fitted, fitted(mySBM))
   expect_equal(mySBM$fitted, predict(mySBM))
-
-  ## correctness
-  expect_lt(rmse(sort(mySBM$connectParam$mean), means), 1e-1)
-  expect_lt(1 - aricode::ARI(mySBM$memberships, mySampler$memberships), 1e-1)
+  expect_equal(predict(mySBM, covarList), fitted(mySBM))
 
 })